write function to get player name to then print to modal.
i know this is probably not the best oop type thinking because the name of a player should be created with a factory. but, the challenges here was that i couldn't reinitialise the game, and there isn't anything like useEffect (like in React) that allows me to repeatedly update the name of each of the players. This is unfortunately the compromise until I refactor this in the future with a bit more experience and knowledge.